requirements.txt
cassandra-driver >= 3.7.1
numpy >= 1.13
##testing
nose>=1.3.7
mock
ccm
